requirements.txt
alabaster==0.7.10
amqp==2.2.2
Babel==2.5.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto==2.48.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.2
Django==1.11.7
django-countries==5.3
django-filter==2.0.0
django-widget-tweaks==1.4.1
docutils==0.14
enum34==1.1.6
eventlet==0.22.0
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
imagesize==1.0.0
Jinja2==2.10
kombu==4.1.0
lxml==4.1.1
MarkupSafe==1.0
numpy==1.13.3
olefile==0.44
packaging==17.1
pandas==0.21.0
Pillow==4.3.0
psycopg2==2.7.3.1
Pygments==2.2.0
pyinstrument==3.0.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scipy==1.0.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
XlsxWriter==1.0.2
xmltodict==0.11.0
regex==2018.8.29